+++ src/Components/Main/index.js
@@ -1,0 +1,2 @@
+import Main from "./Main"
+export default Main;
