+++ src/Components/Sort/index.js
@@ -1,0 +1,2 @@
+import SortBar from "./SortBar";
+export default SortBar;
