+++ src/Containers/MenuContainer/index.js
@@ -1,0 +1,2 @@
+import MenuContainer from "./MenuContainer";
+export default MenuContainer;
